[frontend] migrate and fix Settings
